requirements.txt
attrs==20.2.0
bcrypt==3.2.0
blessings==1.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.1.1
docker==4.3.1
freezegun==1.0.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
jsonschema==3.2.0
packaging==20.4
paramiko==2.7.2
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
pytest-mock==3.4.0
python-dateutil==2.8.1
python-vagrant==0.5.15
PyYAML==5.3.1
requests==2.24.0
scp==0.13.3
six==1.15.0
toml==0.10.1
urllib3==1.25.10
websocket-client==0.57.0
zipp==3.3.1
